refactor(utils): extract MIME helpers into mime.py
Move guess_mime_type, detect_binary_mime_type, parse_gemini_mime_type,
get_default_gemini_mime_type, validate_gemini_mime_type and mime_is_denied into
a new mime.py (depends only on models). utils.py re-exports them; the gemini
response processing still in utils resolves them via that re-import. No behavior
change.
